begin looking up Jesus' words

+++ src/src.rs
@@ -3,7 +3,7 @@ use crate::book::Book;
 #[derive(Debug)]
 pub struct Source {
     pub book: Book,
-    pub chapter: u8,
+    pub chapter: usize,
     // TODO(atec): perhaps some enforcement of verses[0] <= verses[1]
-    pub verses: [u8;2],
+    pub verses: [usize;2],
 }
